export/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.4.0
    # via starlette
asgiref==3.4.1
    # via uvicorn
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via uvicorn
fastapi==0.70.1
    # via -r requirements.in
google-api-core==2.3.2
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.3.3
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.2.1
    # via google-cloud-storage
google-cloud-storage==1.43.0
    # via -r requirements.in
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.1.0
    # via google-cloud-storage
googleapis-common-protos==1.54.0
    # via google-api-core
h11==0.12.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
numpy==1.21.4
    # via
    #   -r requirements.in
    #   pandas
pandas==1.3.5
    # via -r requirements.in
protobuf==3.19.1
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==1.8.2
    # via fastapi
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pandas
pytz==2021.3
    # via pandas
requests==2.26.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-cloud-storage
    #   python-dateutil
sniffio==1.2.0
    # via anyio
starlette==0.16.0
    # via fastapi
typing-extensions==4.0.1
    # via pydantic
urllib3==1.26.7
    # via requests
uvicorn==0.16.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
